+++ Src/Tree/NameTable/NameTablesTypes.h
@@ -17,7 +17,7 @@ struct Name
     void*  localNameTable;
 
     IRRegister reg;    /// < register to address from 
-    size_t     memShift; /// < shift relatively to register
+    int        memShift; /// < shift relatively to register
 };
 
 /// @brief Chosen NAME_TABLE_POISON value for stack
